Rename as VisualInfluxQLEditor
